+++ ng-workspace/projects/OpenWaterFoundation/common/leaflet/map/map.component.spec.ts
@@ -1,0 +1,29 @@
+// import { HttpClientTestingModule } from '@angular/common/http/testing';
+// import { ComponentFixture,
+//           TestBed }                 from '@angular/core/testing';
+
+// import { MapComponent }             from './map.component';
+
+// describe('DialogMapComponent', () => {
+//   let mapComponent: MapComponent;
+//   let fixture: ComponentFixture<MapComponent>;
+
+//   beforeEach(async () => {
+//     await TestBed.configureTestingModule({
+//       declarations: [ MapComponent ],
+//       imports: [ HttpClientTestingModule ]
+//     })
+//     .compileComponents();
+//   });
+
+//   beforeEach(() => {
+//     fixture = TestBed.createComponent(MapComponent);
+//     mapComponent = fixture.debugElement.componentInstance;
+//     fixture.detectChanges();
+//   });
+
+//   it('should create', () => {
+//     expect(mapComponent).toBeTruthy();
+//   });
+  
+// });