+++ ng-workspace/projects/OpenWaterFoundation/common/services/owf-common.service.spec.ts
@@ -1,16 +1,42 @@
-import { TestBed } from '@angular/core/testing';
+import { HttpClientTestingModule,
+          HttpTestingController } from '@angular/common/http/testing';
+import { TestBed }                from '@angular/core/testing';
+import { GRAPH_TEMPLATE }         from './owf-common-test-data';
 
-import { OwfCommonService } from './owf-common.service';
+import { OwfCommonService }       from './owf-common.service';
 
 describe('OwfCommonService', () => {
-  let service: OwfCommonService;
+  let owfCommonService: OwfCommonService,
+      httpTestingController: HttpTestingController;
 
   beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(OwfCommonService);
+    TestBed.configureTestingModule({
+      imports: [ HttpClientTestingModule ]
+    });
+    owfCommonService = TestBed.inject(OwfCommonService);
+    httpTestingController = TestBed.inject(HttpTestingController);
   });
 
-  it('should be created', () => {
-    expect(service).toBeTruthy();
+  describe('when performing asynchronous methods', () => {
+    it('should GET JSON data into an object', () => {
+      expect(owfCommonService).toBeTruthy();
+
+      owfCommonService.getJSONData('assets/app/data-maps/data-ts/streamflow-graph-template.json')
+      .subscribe((data: any) => {
+        expect(data).toBeTruthy('No data returned');
+
+        expect(data.product.properties.Enabled).toContain('True');
+      });
+
+      const req = httpTestingController.expectOne('assets/app/data-maps/data-ts/streamflow-graph-template.json');
+      // console.log('Expect the REST method to be GET.');
+      expect(req.request.method).toEqual('GET');
+      
+      req.flush(GRAPH_TEMPLATE);
+    });
   });
+
+  // describe('when performing synchronous helper methods', () => {
+  // });
+  
 });